+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added this changelog file, closes #8.
 - Added VS Code project files, closes #9.
 - Added a base SerialFillerException class that all serial filler exceptions derive from, closes #10.
+- Added header guard to Crc16Ccitt1021.hpp.
+- Added 'Definitions.hpp' and moved the definitions for 'ByteArray' and 'ByteQueue' to this file.
+
+### Changed
+- Tidied up header includes and improved the method of forward declaration, closes #11.
 
 [Unreleased]: https://github.com/mbedded-ninja/CppTemplate/compare/v1.0.0-beta.1...HEAD
 [v1.0.0-beta.1]: https://github.com/mbedded-ninja/CppTemplate/compare/v0.3.7...v1.0.0-beta.1